refactor(ff-encode): restructure module layout to match ff-decode pattern
Introduce per-type builders (`VideoEncoderBuilder`, `AudioEncoderBuilder`)
in dedicated `builder.rs` files, mirroring the layout already used by
ff-decode. `VideoEncoder::create()` and `AudioEncoder::create()` are now
infallible; validation is deferred to `build()`.

Remove the shared top-level `EncoderBuilder` and the separate
`video/encoder.rs` / `audio/encoder.rs` files that previously split builder
and encoder state across two files.

Closes #151
