+++ crates/ff-encode/src/video/mod.rs
@@ -1,10 +1,10 @@
 //! Video encoding implementation.
 //!
-//! This module provides video encoding functionality with FFmpeg backend.
-//! The implementation is split into public API ([`encoder`]) and internal
+//! This module provides video encoding functionality with an FFmpeg backend.
+//! The implementation is split into public API ([`builder`]) and internal
 //! implementation details ([`encoder_inner`]).
 
-mod encoder;
+pub mod builder;
 mod encoder_inner;
 
-pub use encoder::VideoEncoder;
+pub use builder::{VideoEncoder, VideoEncoderBuilder};
